Fix getBaseVarsList
This function must operate with alias, not with id, since it is called by outside applications who do not know the internal id. What is an id to them is an alias for the autocoder, so neither the input not the output of this function should be an internal id, but, what is called alias internally.